+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project tries to adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- sankey charts
+
+### Changed
+- output format (including support for groups column containing lists)
+
 ## [0.4.1] - 2024-03-05
 ### Added
 - version bumping